requirements.txt
asgiref==3.3.1
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.1.4
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-import-export==2.4.0
django-six==1.0.4
et-xmlfile==1.0.1
jdcal==1.4.1
MarkupPy==1.14
numpy==1.19.4
odfpy==1.4.1
openpyxl==3.0.5
pandas==1.1.5
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.3.1
screen==1.0.1
six==1.15.0
SQLAlchemy==1.3.20
sqlparse==0.4.1
tablib==3.0.0
xlrd==1.2.0
XlsxWriter==1.3.7
xlwt==1.3.0
